backend/requirements.txt
# FastAPI и сервер
fastapi==0.104.1
uvicorn==0.24.0

# База данных
psycopg2-binary==2.9.9

# Авторизация и безопасность
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Конфигурация
pydantic==1.10.13
python-dotenv==1.0.0

# HTTP клиент для тестов
httpx==0.25.2

# Утилиты
itsdangerous==2.1.2